requirements.txt
anthropic>=0.40.0
ollama>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0
requests>=2.31.0
openai>=1.0.0
semgrep>=1.50.0
bandit>=1.7.0
